+++ R/mega.r
@@ -1,17 +1,9 @@
-##' read sequence alignment file from MEGA
-##'
-##'
-##' @title mega_read
-##' @param file mega file
-##' @return BStringSet object
-##' @export
-##' @author Guangchuang Yu
-##' @examples
-##' mega_file <- system.file("extdata/mega/Crab_rRNA.meg", package="seqmagick")
-##' mega_read(mega_file)
+#' @rdname msa-read
+#' @export
+#' @importFrom yulab.utils use_perl
 mega_read <- function(file) {
     x <- readLines(file)
-    start <- grep("^#", x)[2] - 1
+    start <- grep("^#", x, perl=use_perl())[2] - 1
     header <- x[1:start]
     format <- mega_format(header)
 
@@ -20,7 +12,7 @@ mega_read <- function(file) {
         return(res)
     }
 
-    i <- grep("^!Gene=", x)
+    i <- grep("^!Gene=", x, perl=use_perl())
 
     if (length(i) == 0) {
         records <- x[-(1:start)]
@@ -28,9 +20,9 @@ mega_read <- function(file) {
         res <- mega_record(records, format)
         return(res)
     } 
-    genename <- sub("!Gene=", "", x[i])
-    genename <- sub(";", "", genename)
-    genename <- sub("\\s+.*", "", genename)
+    genename <- sub("!Gene=", "", x[i], perl=use_perl())
+    genename <- sub(";", "", genename, perl=use_perl())
+    genename <- sub("\\s+.*", "", genename, perl=use_perl())
 
     begin <- i + 1
     end <- c(i[-1] -1, length(x))
@@ -44,8 +36,8 @@ mega_read <- function(file) {
 }
 
 mega_record <- function(records, format) {
-    records <-  sub("^#", ">", records)
-    records <- gsub("\\s+", "", records)
+    records <-  sub("^#", ">", records, perl=use_perl())
+    records <- gsub("\\s+", "", records, perl=use_perl())
 
     ic <- format['identical']
     type <- format["DataType"]
@@ -74,44 +66,57 @@ mega_record <- function(records, format) {
 }
 
 mega_record_no_format <- function(records, type = NA) {
-    records <- records[records != ""]
-    seqname <- sub("^#(\\S+)\\s+.*", "\\1", records)
-    seqs <- sub("^#\\S+", "", records)
-    seqs <- gsub("\\s+", "", seqs)
+    x <- parse_name_seq(records)
+    seqname <- sub("^#", "", x$seqname, perl=use_perl())
+    seqs <- collapse_seqs(seqname, x$seqs)
+    
+    build_seq_object(seqs, type)
+}
+
+collapse_seqs <- function(seqname, seqs) {
     if (any(duplicated(seqname))) {
         seqs <- vapply(split(seqs, seqname), paste0, collapse = "", FUN.VALUE=character(1))
     } else {
         names(seqs) <- seqname
     }
-
-    if (is.na(type)) {
-        type <- guess_sequence_type(seqs[1])
-    }
-    
-    build_seq_object(seqs, type)
+    return(seqs)
 }
 
+parse_name_seq <- function(records) {
+    records <- records[records != ""]
+    seqname <- sub("^(\\S+)\\s+.*", "\\1", records, perl=use_perl())
+    seqs <- sub("^\\S+", "", records, perl=use_perl())
+    seqs <- gsub("\\s+", "", seqs, perl=use_perl())
+    return(list(seqname = seqname, seqs = seqs))
+}
+
+
 build_seq_object <- function(seq_str, type) {
+    if (is.na(type) || type == "auto") {
+        type <- guess_sequence_type(seq_str[1])
+    }
+
+    type <- toupper(type)
+    
     switch(type,
            DNA = DNAStringSet(seq_str),
            RNA = RNAStringSet(seq_str),
            AA  = AAStringSet(seq_str),
-           Protein  = AAStringSet(seq_str),
+           PROTEIN  = AAStringSet(seq_str),
            UNKNOWN = BStringSet(seq_str)
            )    
 }
 
 mega_format <- function(header) {
-    i <- grep("!Format\\s+DataType=", header)
+    i <- grep("!Format\\s+DataType=", header, perl=use_perl())
     if (length(i) == 0) {
         # not format definition
         return(NULL)
     } 
 
-    f <- gsub("!Format\\s*|;", "", header[i])
+    f <- gsub("!Format\\s*|;", "", header[i], perl=use_perl())
     ff <- strsplit(strsplit(f, " ")[[1]], "=") 
     fm <- do.call(rbind, ff)
     res <- setNames(fm[,2], fm[,1])
     return(res)
 }
-