+++ NEWS.md
@@ -1,6 +1,7 @@
-# seqmagick 0.1.6.001
+# seqmagick 0.1.7
 
-+ `mega_read()` to parse various file types of MEGA sequence alignment (2023-12-19, Tue)
++ `clw_read()` and `sth_read()` to parse CLUSTAL and STOCKHOLM outputs (2023-12-19, Tue)
++ `mega_read()` to parse various file types of MEGA sequence alignment (2023-12-18, Mon)
 
 # seqmagick 0.1.6